HomusTrainer/requirements.txt
# Use conda install ... to install these
matplotlib
pillow
graphviz

# Use pip install ... to install these
keras == 2.0.4
tensorflow == 1.1.0
# tensorflow-gpu == 1.1.0 # Alternatively if available.
codecov
pytest
pytest-cov
scipy
h5py
scikit-learn